translate/requirements.txt
Flask==0.12.2
google-cloud-translate==1.2.0
gunicorn==19.7.1